Делает рефакторинг модуля big picture
