+++ js/big-picture.js
@@ -1,17 +1,9 @@
 import { isEscapeKey } from './util.js';
-
-const COMMENTS_NUMBER = 5;
+import { initComments } from './comments.js';
 
 const bodyElement = document.querySelector('body');
 const bigPictureElement = document.querySelector('.big-picture');
 const closeBigPictureElement = document.querySelector('.big-picture__cancel');
-const commentCountElement = document.querySelector('.social__comment-count');
-const commentsLoaderElement = document.querySelector('.comments-loader');
-const commentsContainerElement = bigPictureElement.querySelector('.big-picture__social .social__comments');
-const commentTemplate = document.querySelector('#comment').content.querySelector('.social__comment');
-
-let currentComments;
-let commentsShown = 0;
 
 const onDocumentKeydown = (evt) => {
   if (isEscapeKey(evt)) {
@@ -28,35 +20,6 @@ const renderBigPicture = ({ url, description, likes, comments }) => {
   bigPictureElement.querySelector('.social__caption').textContent = description;
 };
 
-const renderComments = () => {
-  const newCommentsShown = Math.min(currentComments.length, commentsShown + COMMENTS_NUMBER);
-
-  const fragment = document.createDocumentFragment();
-
-  const pieceCurrentComments = currentComments.slice(commentsShown, newCommentsShown);
-
-  pieceCurrentComments.forEach(({ avatar, name, message }) => {
-    const commentElement = commentTemplate.cloneNode(true);
-    const commentImage = commentElement.querySelector('.social__picture');
-    commentImage.src = avatar;
-    commentImage.alt = name;
-    const commentText = commentElement.querySelector('.social__text');
-    commentText.textContent = message;
-    fragment.appendChild(commentElement);
-  });
-  commentsShown = newCommentsShown;
-
-  if (commentsShown >= currentComments.length) {
-    commentsLoaderElement.classList.add('hidden');
-  } else {
-    commentsLoaderElement.classList.remove('hidden');
-  }
-
-  commentsContainerElement.appendChild(fragment);
-
-  commentCountElement.innerHTML = `${commentsShown} из <span class="comments-count">${currentComments.length}</span> комментариев`;
-};
-
 const openBigPicture = (info) => {
   bigPictureElement.classList.remove('hidden');
   bodyElement.classList.add('modal-open');
@@ -64,21 +27,14 @@ const openBigPicture = (info) => {
 
   renderBigPicture(info);
 
-  commentsContainerElement.innerHTML = '';
-  currentComments = info.comments;
-  renderComments();
+  initComments(info.comments);
 };
-
-commentsLoaderElement.addEventListener('click', () => {
-  renderComments();
-});
 
 closeBigPictureElement.addEventListener('click', () => {
   bigPictureElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
-  commentsShown = 0;
 
   document.removeEventListener('keydown', onDocumentKeydown);
 });
 
-export { renderBigPicture, openBigPicture };
+export { openBigPicture };
